requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-ollama
langgraph
langgraph-checkpoint-postgres
langgraph-checkpoint-sqlite
rich
google-api-python-client
google-auth-httplib2
google-auth-oauthlib